requirements.txt
# Parabola RM Builder Requirements

# Core dependencies
pyyaml>=5.1
argparse>=1.4.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.10.0
black>=20.8b1
flake8>=3.8.0
mypy>=0.800